split out validations into rules, move some logic out of handlers but test broken
