+++ pkg/handlers/primeapi/mto_shipment.go
@@ -2,7 +2,6 @@ package primeapi
 
 import (
 	"fmt"
-	"time"
 
 	"github.com/go-openapi/runtime/middleware"
 	"github.com/gobuffalo/validate/v3"
@@ -203,137 +202,6 @@ func (h DeleteMTOShipmentHandler) Handle(params mtoshipmentops.DeleteMTOShipment
 
 			return mtoshipmentops.NewDeleteMTOShipmentNoContent(), nil
 		})
-}
-
-// This function checks Prime specific validations on the model
-// It expects dbShipment to represent what's in the db and mtoShipment to represent the requested update
-// It updates mtoShipment accordingly if there are dependent updates like requiredDeliveryDate
-// On completion it either returns a list of errors or an updated MTOShipment that should be stored to the database.
-func (h UpdateMTOShipmentHandler) checkPrimeValidationsOnModel(appCtx appcontext.AppContext, mtoShipment *models.MTOShipment, dbShipment *models.MTOShipment) (*models.MTOShipment, *validate.Errors) {
-	verrs := validate.NewErrors()
-
-	// Prime cannot edit the customer's requestedPickupDate
-	if mtoShipment.RequestedPickupDate != nil {
-		requestedPickupDate := mtoShipment.RequestedPickupDate
-		if !requestedPickupDate.Equal(*dbShipment.RequestedPickupDate) {
-			verrs.Add("requestedPickupDate", "must match what customer has requested")
-		}
-		mtoShipment.RequestedPickupDate = requestedPickupDate
-	}
-
-	// Get the latest scheduled pickup date as it's needed to calculate the update range for PrimeEstimatedWeight
-	// And the RDD
-	latestSchedPickupDate := dbShipment.ScheduledPickupDate
-	if mtoShipment.ScheduledPickupDate != nil {
-		latestSchedPickupDate = mtoShipment.ScheduledPickupDate
-	}
-
-	// Prime can update the estimated weight once within a set period of time
-	// If it's expired, they can no longer update it.
-	latestEstimatedWeight := dbShipment.PrimeEstimatedWeight
-	if mtoShipment.PrimeEstimatedWeight != nil {
-		if dbShipment.PrimeEstimatedWeight != nil {
-			verrs.Add("primeEstimatedWeight", "cannot be updated after initial estimation")
-		}
-		// Validate if we are in the allowed period of time
-		now := time.Now()
-		if latestSchedPickupDate != nil {
-			err := validatePrimeEstimatedWeightRecordedDate(now, *latestSchedPickupDate)
-			if err != nil {
-				verrs.Add("primeEstimatedWeight", "the time period for updating the estimated weight for a shipment has expired, please contact the TOO directly to request updates to this shipment’s estimated weight")
-				verrs.Add("primeEstimatedWeight", err.Error())
-			}
-		}
-		// If they can update it, it will be the latestEstimatedWeight (needed for RDD calc)
-		// And we also record the date at which it happened
-		latestEstimatedWeight = mtoShipment.PrimeEstimatedWeight
-		mtoShipment.PrimeEstimatedWeightRecordedDate = &now
-	}
-
-	// Prime cannot update or add agents with this endpoint, so this should always be empty
-	if len(mtoShipment.MTOAgents) > 0 {
-		if len(dbShipment.MTOAgents) < len(mtoShipment.MTOAgents) {
-			verrs.Add("agents", "cannot add or update MTO agents to a shipment")
-		}
-	}
-
-	// Prime can create a new address, but cannot update it.
-	// So if address exists, return an error. But also set the local pointer to nil, so we don't recalculate requiredDeliveryDate
-	var latestPickupAddress *models.Address
-	var latestDestinationAddress *models.Address
-
-	switch dbShipment.ShipmentType {
-	case models.MTOShipmentTypeHHGIntoNTSDom:
-		if dbShipment.StorageFacility == nil {
-			// latestDestinationAddress is only used for calculating RDD.
-			// We don't want to block an update because we're missing info to calculate RDD
-			break
-		}
-		latestPickupAddress = dbShipment.PickupAddress
-		latestDestinationAddress = &dbShipment.StorageFacility.Address
-	case models.MTOShipmentTypeHHGOutOfNTSDom:
-		if dbShipment.StorageFacility == nil {
-			// latestPickupAddress is only used for calculating RDD.
-			// We don't want to block an update because we're missing info to calculate RDD
-			break
-		}
-		latestPickupAddress = &dbShipment.StorageFacility.Address
-		latestDestinationAddress = dbShipment.DestinationAddress
-	default:
-		latestPickupAddress = dbShipment.PickupAddress
-		latestDestinationAddress = dbShipment.DestinationAddress
-	}
-	// We also track the latestPickupAddress for the RDD calculation
-	if dbShipment.PickupAddress != nil && mtoShipment.PickupAddress != nil { // If both are populated, return error
-		verrs.Add("pickupAddress", "the pickup address already exists and cannot be updated with this endpoint")
-	} else if mtoShipment.PickupAddress != nil { // If only the update has an address, that's the latest address
-		latestPickupAddress = mtoShipment.PickupAddress
-	}
-	if dbShipment.DestinationAddress != nil && mtoShipment.DestinationAddress != nil {
-		verrs.Add("destinationAddress", "the destination address already exists and cannot be updated with this endpoint")
-	} else if mtoShipment.DestinationAddress != nil {
-		latestDestinationAddress = mtoShipment.DestinationAddress
-	}
-
-	// For secondary addresses we do the same, but don't have to track the latest values for RDD
-	if dbShipment.SecondaryPickupAddress != nil && mtoShipment.SecondaryPickupAddress != nil { // If both are populated, return error
-		verrs.Add("secondaryPickupAddress", "the secondary pickup address already exists and cannot be updated with this endpoint")
-	}
-	if dbShipment.SecondaryDeliveryAddress != nil && mtoShipment.SecondaryDeliveryAddress != nil {
-		verrs.Add("secondaryDeliveryAddress", "the secondary delivery address already exists and cannot be updated with this endpoint")
-	}
-
-	// If we have all the data, calculate RDD
-	if latestSchedPickupDate != nil && (latestEstimatedWeight != nil || (dbShipment.ShipmentType == models.MTOShipmentTypeHHGOutOfNTSDom &&
-		dbShipment.NTSRecordedWeight != nil)) && latestPickupAddress != nil && latestDestinationAddress != nil {
-		weight := latestEstimatedWeight
-		if dbShipment.ShipmentType == models.MTOShipmentTypeHHGOutOfNTSDom && dbShipment.NTSRecordedWeight != nil {
-			weight = dbShipment.NTSRecordedWeight
-		}
-		requiredDeliveryDate, err := mtoshipment.CalculateRequiredDeliveryDate(appCtx, h.Planner(), *latestPickupAddress,
-			*latestDestinationAddress, *latestSchedPickupDate, weight.Int())
-		if err != nil {
-			verrs.Add("requiredDeliveryDate", err.Error())
-		}
-		mtoShipment.RequiredDeliveryDate = requiredDeliveryDate
-	}
-
-	return mtoShipment, verrs
-}
-
-func validatePrimeEstimatedWeightRecordedDate(estimatedWeightRecordedDate time.Time, scheduledPickupDate time.Time) error {
-	recordedYear, recordedMonth, recordedDate := estimatedWeightRecordedDate.Date()
-	scheduledYear, scheduledMonth, scheduledDate := scheduledPickupDate.Date()
-
-	if estimatedWeightRecordedDate.Before(scheduledPickupDate) {
-		return nil
-	}
-
-	if recordedYear == scheduledYear && recordedMonth == scheduledMonth && recordedDate == scheduledDate {
-		return nil
-	}
-
-	return apperror.InvalidInputError{}
 }
 
 // UpdateMTOShipmentStatusHandler is the handler to update MTO Shipments' status
